+++ scripts/browser.py
@@ -1,3 +1,0 @@
-from _compat import alias
-
-alias("browser", __name__)