+++ scripts/cache_validity.py
@@ -1,3 +1,0 @@
-from _compat import alias
-
-alias("cache_validity", __name__)