+++ scripts/captcha_selenium.py
@@ -1,3 +1,0 @@
-from _compat import alias
-
-alias("captcha_selenium", __name__)