+++ scripts/config.py
@@ -1,3 +1,0 @@
-from _compat import alias
-
-alias("config", __name__)