+++ scripts/const.py
@@ -1,3 +1,0 @@
-from _compat import alias
-
-alias("const", __name__)