+++ scripts/ha_mapping.py
@@ -1,3 +1,0 @@
-from _compat import alias
-
-alias("ha_mapping", __name__)