+++ scripts/login.py
@@ -1,3 +1,0 @@
-from _compat import alias
-
-alias("login", __name__)