+++ scripts/model.py
@@ -1,3 +1,0 @@
-from _compat import alias
-
-alias("model", __name__)