+++ scripts/notify.py
@@ -1,3 +1,0 @@
-from _compat import alias
-
-alias("notify", __name__)