+++ scripts/redact.py
@@ -1,3 +1,0 @@
-from _compat import alias
-
-alias("redact", __name__)