+++ scripts/session.py
@@ -1,3 +1,0 @@
-from _compat import alias
-
-alias("session", __name__)