+++ scripts/store.py
@@ -1,3 +1,0 @@
-from _compat import alias
-
-alias("store", __name__)